refactor: extract parseOptions to shared argument parser utility
Created reusable argument parsing utilities in utils/arg-parser.ts:

- Moved parseOptions function from mcp-add-utils.ts to shared location
- Enhanced with comprehensive JSDoc documentation and examples
- Added ParsedArguments interface for better type safety
- Added reconstructArgs utility function for debugging/testing
- Improved code reusability across different command modules

The parseOptions function handles:
- Regular arguments (non-option arguments)
- --key=value style options
- --flag style boolean flags

This makes the argument parsing logic available for other commands
that need to parse command-line options in a consistent way.
